modified:   main/admin.py 	new file:   main/migrations/0003_rename_huella_objetivos_huella_and_more.py 	new file:   main/migrations/0004_acciones_delete_objetivos.py 	new file:   main/migrations/0005_alter_acciones_huella.py 	modified:   main/models.py

+++ main/models.py
@@ -6,7 +6,7 @@ class Acciones(models.Model):
     Accion = models.CharField(max_length=50)
     Categoria = models.CharField(max_length=50)
     Acumulable = models.BooleanField()
-    Huella = models.IntegerField()
+    Huella = models.DecimalField(max_digits=5,decimal_places=2)
 
 class Usuarios(models.Model):
     Usuario = models.CharField(max_length=20)
